ConfigManager to its own project
