+++ src/LogExpert.Tests/JSONSaveTest.cs
@@ -1,11 +1,9 @@
-﻿using LogExpert.Config;
+using LogExpert.Configuration;
 using LogExpert.Core.Config;
 
 using Newtonsoft.Json;
 
 using NUnit.Framework;
-
-using System.IO;
 
 namespace LogExpert.Tests;
 
@@ -13,7 +11,7 @@ namespace LogExpert.Tests;
 public class JSONSaveTest
 {
     [Test(Author = "Hirogen", Description = "Save Options as JSON and Check if the written file can be cast again into the settings object")]
-    public void SaveOptionsAsJSON()
+    public void SaveOptionsAsJSON ()
     {
         ConfigManager.Instance.Settings.AlwaysOnTop = true;
         ConfigManager.Instance.Save(SettingsFlags.All);
@@ -21,14 +19,14 @@ public class JSONSaveTest
         var settingsFile = configDir + "\\settings.json";
 
         Settings settings = null;
-        
+
         Assert.DoesNotThrow(CastSettings);
         Assert.That(settings, Is.Not.Null);
         Assert.That(settings.AlwaysOnTop, Is.True);
 
         ConfigManager.Instance.Settings.AlwaysOnTop = false;
         ConfigManager.Instance.Save(SettingsFlags.All);
-        
+
         settings = null;
         Assert.DoesNotThrow(CastSettings);
 
@@ -36,7 +34,7 @@ public class JSONSaveTest
         Assert.That(settings.AlwaysOnTop, Is.False);
 
 
-        void CastSettings()
+        void CastSettings ()
         {
             settings = JsonConvert.DeserializeObject<Settings>(File.ReadAllText(settingsFile));
         }
