+++ src/LogExpert/LogExpert.csproj
@@ -30,6 +30,7 @@
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\ColumnizerLib\ColumnizerLib.csproj" />
+    <ProjectReference Include="..\LogExpert.Configuration\LogExpert.Configuration.csproj" />
     <ProjectReference Include="..\LogExpert.UI\LogExpert.UI.csproj" />
     <ProjectReference Include="..\PluginRegistry\LogExpert.PluginRegistry.csproj" />
   </ItemGroup>
