+++ src/LogExpert/Program.cs
@@ -11,7 +11,7 @@ using System.Windows.Forms;
 
 using LogExpert.Classes;
 using LogExpert.Classes.CommandLine;
-using LogExpert.Config;
+using LogExpert.Configuration;
 using LogExpert.Core.Classes.IPC;
 using LogExpert.Core.Config;
 using LogExpert.Dialogs;
@@ -55,6 +55,9 @@ internal static class Program
         Application.EnableVisualStyles();
         Application.SetUnhandledExceptionMode(UnhandledExceptionMode.CatchException);
 
+        // Initialize ConfigManager with application-specific paths and screen information
+        ConfigManager.Instance.Initialize(Application.StartupPath, SystemInformation.VirtualScreen);
+
         _logger.Info(CultureInfo.InvariantCulture, $"\r\n============================================================================\r\nLogExpert {Assembly.GetExecutingAssembly().GetName().Version.ToString(3)} started.\r\n============================================================================");
 
         CancellationTokenSource cts = new();
@@ -95,9 +98,9 @@ internal static class Program
                 }
             }
 
-            SetCulture();
-
             _ = PluginRegistry.PluginRegistry.Create(ConfigManager.Instance.ConfigDir, ConfigManager.Instance.Settings.Preferences.PollingInterval);
+
+            SetCulture();
 
             ColumnizerLib.Column.SetMaxDisplayLength(ConfigManager.Instance.Settings.Preferences.MaxDisplayLength);
